+++ bubbletea03-my-app/src/page/CategorySelect.js
@@ -1,4 +1,4 @@
-export default function Home() {    
+export default function CategorySelect() {    
     return (
         <Container>
