+++ bubbletea03-my-app/src/page/ProductPage.js
@@ -1,14 +1,16 @@
-import { useParams } from 'react-router-dom';
+import { useParams, useNavigate } from 'react-router-dom';
 import UserContent from '../component/product-page/UserContent';
 import UserItemList from '../component/product-page/UserItemList';
 import styled from 'styled-components';
 import itemData from '../data/ItemData';
-import TopNav from './../component/product-page/TopNav';
 import BottomNav from './../component/product-page/BottomNav';
 import { useRecoilState } from 'recoil';
 import { testState } from '../RecoilStates';
+import TopNav from '../component/TopNav';
+import { Link } from 'react-router-dom';
 
 export default function ProductPage() {
+    const navigate = useNavigate();
     const params = useParams();
     const [ item ] = itemData.filter(item => item.id == params.itemId);
     const [test, setTest] = useRecoilState(testState);
@@ -16,7 +18,14 @@ export default function ProductPage() {
     window.scrollTo(0, 0);
     return (
         <Container>
-            <TopNav/>
+            <TopNav left={<>
+                <button onClick={() => {navigate(-1)}}>
+                    <img alt="왼쪽 화살표" src="/icon/arrow.png"/>
+                </button>
+                <Link to="/">
+                    <img alt="홈 아이콘" src="/icon/home.png"/>
+                </Link>
+            </>}/>
             <BottomNav/>
             <div onClick={() => {setTest("ㅇㅇㅋㅋ")}}>ddd{test}</div>
             <TopImage>
